Delete Cargo.lock
Rust ignores Cargo.lock files in dependent libraries.  That means that
committing a Cargo.lock file to a library has no effect on the binaries
that get built from that library.  The only affect it can have is during
development and during CI.  So you shouldn't use them with a crate like
this.